Rename BUILD.bazel file to BUILD
Bazel is very particular about what name its BUILD file has. When using
the `git_repository` Bazel command, there is no way to tell it what name
to use for a BUILD file. In fact, it's very complicated to deal with.
Ultimately, one would have to download the repo, rename the file, then
import it into Bazel as a local repo.

If we just rename the BUILD.bazel file to BUILD, then Bazel will hum
along happily. I think that, if we want to offer a Bazel build (which is
a good idea), we should probably make using it as smooth as possible.